Crate miden_debug_types

Re-exports

pub use serde_spanned;

Structs

ByteIndex

An index representing the offset in bytes from the start of a source file

ByteOffset

An offset in bytes relative to some ByteIndex

ColumnIndex

A zero-indexed column number

ColumnNumber

A one-indexed column number

DefaultSourceManager

FileLineCol

A FileLineCol represents traditional file/line/column information for use in rendering.

LineIndex

A zero-indexed line number

LineNumber

A one-indexed line number

Location

A Location represents file and span information for portability across source managers

Position

Position in a text document expressed as zero-based line and character offset.

Selection

A range in a text document expressed as (zero-based) start and end positions.

SourceContent

Represents key information about a source file and its content:

SourceFile

A SourceFile represents a single file stored in a super::SourceManager

SourceFileRef

A reference to a specific spanned region of a SourceFile, that provides access to the actual SourceFile, but scoped to the span it was created with.

SourceId

A SourceId represents the index/identifier associated with a unique source file in a SourceManager implementation.

SourceSpan

This represents a span of bytes in a Miden Assembly source file.

Span

This type is used to wrap any T with a SourceSpan, and is typically used when it is not convenient to add a SourceSpan to the type - most commonly because we don’t control the type.

Uri

A URI reference that specifies the location of a source file, whether on disk, on the network, or elsewhere.

Enums

SourceContentUpdateError

SourceLanguage

Traits

SourceManager

SourceManagerExt

SourceManagerSync

SourceManagerSync is a marker trait for SourceManager implementations that are also Send + Sync, and is automatically implemented for any SourceManager that meets those requirements.

Spanned

This trait should be implemented for any type that has an associated SourceSpan.