miden_debug_types

Trait SourceManagerSync

pub trait SourceManagerSync:
    SourceManager
    + Send
    + Sync { }

SourceManagerSync is a marker trait for SourceManager implementations that are also Send + Sync, and is automatically implemented for any SourceManager that meets those requirements.

SourceManager is a supertrait of SourceManagerSync, so you may use instances of the SourceManagerSync where the SourceManager is required, either implicitly or via explicit downcasting, e.g. Arc<dyn SourceManagerSync> as Arc<dyn SourceManager>.